pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expo-prediction"
version = "0.1.0"
description = "ExPO: Exposure-Response Neural Operator for Gene Expression Imputation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MLBC Lab", email = "info@mlbc-lab.org"}
]
keywords = ["gene-expression", "neural-operators", "drug-discovery", "machine-learning", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "scipy>=1.11",
    "tqdm>=4.66",
    "torch>=2.1",
    "transformers>=4.40",
    "rdkit-pypi>=2022.9.5",
    "pyyaml>=6.0",
    "pytest>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.18",
]

[project.urls]
"Homepage" = "https://github.com/MLBC-lab/ExPO"
"Bug Tracker" = "https://github.com/MLBC-lab/ExPO/issues"
"Documentation" = "https://github.com/MLBC-lab/ExPO#readme"
"Source Code" = "https://github.com/MLBC-lab/ExPO"

[project.scripts]
expo-train = "expo.cli:train_main"
expo-eval = "expo.cli:eval_main"
expo-demo = "expo.cli:demo_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["expo*"]

[tool.setuptools.package-data]
expo = ["*.yaml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true